Export commands
The ServerCommand and Command structures are currently needed (or at
least sensible given the goal of providing the bot with all information
available to the client) but are very cludgy in this context.  The Bot
module should probably do a better job of translating between internal
structures and those used by Ninjas.